Rename SortedSetPropertyFactory.CodeGenerator -> SortedSetProperty
SortedSetPropertyFactory is now SortedSetProperty.Factory
